automation: Require coverage 5
Looks like recent failures in coverage are caused by incompatibility
between pytest and pytest-cov and older coverage[1]. We don't specify
coverage version, so we get some old version that happened to be
installed on the slaves.

Upgrade to coverage >= 5. Since this version is not packaged yet for
Centos, and this is developer tool and not runtime requirement, install
coverage using pip. This will make it easier to use latest pytest and
pytest-cov.

Ideally this would be installed by tox in a virtual environment, but we
use coverage outside of tox to create a combined coverage report. For
the CI environment, using pip is fine.

[1] https://stackoverflow.com/questions/59439831/coveralls-unicodedecodeerror

Change-Id: I5b252fb1b5fe266d25b457b42145b176ffba0e80
Signed-off-by: Nir Soffer <nsoffer@redhat.com>