livemerge: Make the return value more clear
Rename meaningless (and ugly) jobsRet to tracked_jobs. After spending
some time reading this code, I'm pretty sure that I understand how it
works, and this makes this code more clear.

In the unlikely case when we don't find the job drive, we don't report
the job info. This looks like a bug that we need to think about later.

Change-Id: I9209c9ffc63605f6f1d0142311b4dd8f971a894a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>