supervdsm: enable compression for NVRAM data
Compressing the data with bzip2 makes the final (base64 encoded) string
around 7x smaller.

Instead of enabling the compression only for NVRAM we change the
default to always have compression. In case of TPM we explicitly disable
the compression for backward-compatibility. Ideally we would like to
enable compression also for TPM in the future.

Change-Id: I1b52cea797760649f81a74db77317f6c67d61065
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>