automation: Log installed packages versions
We are fighting with CI and repos trying get decent qemu-kvm and
qemu-img so we don't xfail and skip many import tests, but it is very
hard to understand what's going on since we don't have enough visibility
on the slave packages.

Log installed qemu-kvm and qemu-img version on the start of the build to
make this easy.

Change-Id: If24d0bdbaca0a7d94e4ff3cc81e06fbc6bc6d414
Signed-off-by: Nir Soffer <nsoffer@redhat.com>