caps: add capability for specifying CD via PDIV
In previous patches was added support for specifying CD via PDIV, which
is the preferred way who to specify the image. To let engine know that
host is able to work with images specified by PDIV, add new capability
`cd_change_pdiv` simple set to True if host support this functionality.

Change-Id: I45ddd56d3f03d2c254e21479d2ac4f2c660b89a0
Bug-Url: https://bugzilla.redhat.com/1589763
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>