add_bitmap.py: perform operation validation after chain prepared
In order to validate that the bitmap doesn't exit in the volume
chain there is a use with 'qemu-img info' command.
For block-based volumes, the command will fail because the
validation is done before the preparation of the chain.

This is now fixed by moving the validation to occur only after
the chain prepared.

Change-Id: If907ab60c63fff8a63f5a7d350f47d3547b21db9
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>