virt: filedata: use long options for tar, add --verbose
For clarity short option names were replaced by long names.

While at it, adding --verbose option to get the list of files. The
behavior is slightly different for create and for extract though. When
extracting tar the file list goes to stdout and thus is only logged in
case of errors. On the other hand, when creating a tar, the file list
goes to stderr (because compressed data go to stdout) and thus is logged
always.

Change-Id: Ie584892a14bfbfad3270a97b20f9d91ae9974962
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>