imagetickets: refactor request() to private function
Parsing of response content has been moved to caller
functions and request() should be marked private.

This requires changes to tests as well and removing
direct testing of request() and replacing with
public functions.

By testing only the public functions, we never have
to change the tests when changing the implementation.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1858956
Change-Id: If9e396ca8e472a79efcef7923af315c9ad70b5e8