virt: Add VM Reset call
Add a way to reset a VM next to reboot/shutdown.
This as you sometimes want to quickly reset a VM that has crashed/is OOM
for example. The reset call is much quicker than a shutdown/start cycle.

When we reset the VM we set the status to:
vmstatus.REBOOT_IN_PROGRESS

Libvirt will also send a VIR_DOMAIN_EVENT_ID_REBOOT, and this is already
handled within VDSM.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1927718
Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>
Change-Id: I8cbe479353de6ac33fa5506e9c3d4469bbc49187