hba: Use commands.wait_async()
Use commands.wait_saync() instead of zombiereaper to wait for slow
fc-scan after a timeout.

This does not solve issue of starting a new scan before the first scan
finished. The new scan is likely to block in the kernel until the
previous scan completed.

This removes 1/4 uses of zombiereaper.

Change-Id: I0e457261bc627f317bf28e92554edb01aad54c4c
Bug-Url: https://bugzilla.redhat.com/1926589
Signed-off-by: Nir Soffer <nsoffer@redhat.com>