iscsiadm: Use commands.wait_async()
Use commands.wait_saync() instead of zombiereaper to wait for slow
fc-scan after a timeout.

This does not solve issue of starting a new scan before the first scan
finished. The new scan is likely to block in the kernel until the
previous scan completed.

This command run with sudo, so it is not strictly required to fix the fd
leak in supervdsm, but we can have the same issue in vdsm, so we want to
remove all usage of zombiereaper.

This removes 1/3 uses of zombiereaper.

Change-Id: I4fe32c216cc111160181885f2cb4fbcf1f3c08f6
Bug-Url: https://bugzilla.redhat.com/1926589
Signed-off-by: Nir Soffer <nsoffer@redhat.com>