supervdsm_server: Simplify communication with child
The current code is confusing; multiprocessing.Pipe() returns 2
connections, not pipes. Both the parent and child read and write from
both ends of the pipe, which seems pointless.

Simplify the communication with the child by using non-duplex pipe. The
child gets the write end and sends the result. The parent polls the read
end and receives the result.

Change-Id: Iaaf4f49cfaa4ab90bde6b66a2b6b9426a6bb3998
Bug-Url: https://bugzilla.redhat.com/1926589
Signed-off-by: Nir Soffer <nsoffer@redhat.com>