supervdsm_server: Simplify error handling
Error handling in _runAs was more complex than needed. Simplify the flow
using try-finally, and minimize the usage of asynchronous reaping by
terminating the child gracefully, and waiting for it one second before
killing it and passing it to zombie reaper.

We used to check if the process exists after sending a kill signal. This
should never be needed if we check that the process is a live. The
process cannot disappear before we join it.

Change-Id: Ie3b3698f5488ff5a5dc82ff23db63dd66ba164b2
Bug-Url: https://bugzilla.redhat.com/1926589
Signed-off-by: Nir Soffer <nsoffer@redhat.com>