v2v: Always wait for virt-v2v
Previously we waited only 30 seconds after virt-v2v finished, or when
aborting the process, and we pass the pids to zombiereaper. We can
replace zombiereaper with commands.wait_async(), but I don't think we
should keep live virt-v2v process accessing storage like that.

This removes the last usage of zombiereaper from vdsm.

Change-Id: Icfd8dd882f7995c57b2136afb05e2e255d7b3d87
Bug-Url: https://bugzilla.redhat.com/1926589
Signed-off-by: Nir Soffer <nsoffer@redhat.com>