virt: v2v: do not use direct backend for most import methods
The only problematic input is still Xen. We have to rely on direct
backend until all the issues are solved.

Instances of guestfs are started in session libvirt and not in system
libvirt. On the bright side those VMs cannot confuse VDSM, but it may
be unexpected surprise for those debuging issues with import.

Bug-Url: https://bugzilla.redhat.com/1370055
Change-Id: I953aab12004c92fa8f238dbe3790e9ece385af99
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>