tests: Fix headers in FakeResponse
Fix FakeResponse to add content-type header with default content type
that works for most tests.

Passing headers to FakeResponse is not very useful since we must
replicate the code to create correct content-length. It is much easier
to modify the good headers created by the fake response after.

We had test for 204 "No content" response with and without
content-length. Since new imageio never return content-length in this
case, there is no point to continue to test the old behavior. Move the
default behavior to the FakeResponse.

Change-Id: I4c64de72d01dc33921455633c58aecd8c0fc4999
Signed-off-by: Nir Soffer <nsoffer@redhat.com>