livemerge: Replace job dict with a Job class
To minimize the change and make it safer and easier to review I kept the
dict key name as class attribute, and made all class attributes
writable.

One advantage is removing some constant keys. They are now read only
properties, so we don't need to pass them around or serialize them.

To implement serialization for vm metadata, the class supports to_dict()
and from_dict methods.

This change makes the code little bit nicer, replacing dict access
(job["baseVolume"]) with attribute access (job.baseVolume).

Live info is not part of the class yet, and it also does not know how to
report job status. This will be added in followup patches.

The change is based on a much bigger patch from Amit:
https://gerrit.ovirt.org/c/vdsm/+/112319

Change-Id: I04761864de30b425a1d0572a9e43ed336f742f37
Signed-off-by: Nir Soffer <nsoffer@redhat.com>