livemerge: Unify base and top volume naming
We used "baseVolUUID" and "topVolUUID" in the API level, and stored
"baseVolume" and "topVolume" in the job metadata.

Unify the terms and use "base" and "top" everywhere. These are the terms
used by libvirt for blockCommit, and using the same terms and shorter
names makes the code nicer to work with. These are also the terms used
in most of the logs.

Change-Id: I83676fb20106103acdfb17686120463e7fbf6569
Signed-off-by: Nir Soffer <nsoffer@redhat.com>