livemerge: Keep job live info in Job
Previously we got live info from libvirt in query_jobs() and while
querying the jobs we built a dict for reporting job status to engine.
This makes the code more complicated and hard to understand.

We want to separate querying the job from reporting status. The first
step is to keep the live info in the Job object, and provide a info()
method returning the info expected by engine.

Change-Id: I29ce54b6889d434f51bb7839654054a7ba678cd0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>