livemerge: Modernize error handling
Replace error handler like:

   log.error("reason...")
   return response.error("mergeErr")

With raising the corresponding error with the relevant context:

   raise exception.MergeFailed("reason...", job=job_id)

This is converted to proper response in API.py.

Remove the return value for normal case since it is also generated by
API.py.

This will make it easy to split the huge merge() method to smaller
functions without passing around return values.

The first example is _can_merge_info() converted to
_validate_base_size() so we can raise the public exception inside it and
simplify error handling in merge().

Change-Id: I23be01e42fa8ef5bdd5039b3c7581204533d863f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>