coverage: Fix coverage collection
We've had all the parts necessary to run vdsm with coverage
enabled since a long time. The 'coverage' library was always
specific, that it needs its registered 'atexit' handlers to run,
for the coverage data dump to be successful. At some point in time
the coverage setup stopped working. It's hard to tell when exactly,
but my guess is it simply never worked on py3.

After some analysis it turned out that vdsm's and supervdsm's
custom SIGTERM handlers cause the registered 'atexit' functions
not to be called.  This makes the coverage report work only
for the spawned sub-processes, but not the main process.
The effect is that a basic suite run reports ~35% coverage [2]
on vdsm's codebase (with some obvious paths missing like most
of the 'virt.vm' module). With this patch it's ~58% [3].

Unfortunately there's no other way for the 'atexit' handlers
to execute other that calling the private 'atexit._run_exitfuncs'
function. That's why the code is called only when the coverage
is enabled in the config.

With this patch and some pending patches to OST [4] the vdsm coverage
should work properly again soon.

[1] https://coverage.readthedocs.io/en/coverage-5.4/subprocess.html?highlight=atexit#signal-handlers-and-atexit
[2] https://jenkins.ovirt.org/view/oVirt%20system%20tests/job/ovirt-system-tests_manual/7782/artifact/exported-artifacts/coverage/html/index.html
[3] https://jenkins.ovirt.org/job/ovirt-system-tests_manual/7784/artifact/exported-artifacts/coverage/html/index.html
[4] https://gerrit.ovirt.org/#/q/topic:ost-vdsm-coverage+(status:open+OR+status:merged)

Change-Id: Ic28863cfca971f741390265f8ff2ef01114c77f8
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>