livemerge: Create job object early in merge
Extracting helpers from merge() with the current code is hard, since we
have too many variables that need to be passed to the helpers.

Solve this issue by creating a job object early as possible and use this
object instead of referring directly to base, top, bandwidth, and
job_id.

To make job more useful as a way to pass job details, the bandwidth was
added to the job object. Like other attributes it is persisted now in
the vm metadata.

Change-Id: Ifd248ff51d9c686dc5d70a82644914f956868e3e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>