livemerge: Extract _start_commit() helper
Now that we reference only job and drive objects, it is very easy to
extract the code to start a commit block job to a helper. In this patch
it just makes the flow easier to follow, but in future patch we will
start the commit only after extending the base volume was completed.

Change-Id: I0cb62f3e9bd771db3749ecc7f1ae2c6d062ebccc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>