livemerge: Extract _create_job() helper from merge()
Creating a job is pretty verbose since we have to construct a dict from
the drive. It would be nice if engine was sending a useful job
description but it does not, so we have to create it here.

With this change merge() fits nicely in one laptop screen and is very
easy to follow and modify.

Change-Id: I07723dd355a3d78c3d34290fd9a04106e5c0866a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>