blockSD: omit missing keys when dumping sd
When block storage domain dumps information via
.dump() public function it tries to get metadata
from slots - directly from "metadata" lv and if
image, parent or md information is missing it
tries to use lv tags instead. When those are
missing as well we used to return None which is
not correct.

The convention is to leave out the missing
values/keys from the output completely when
lookup fails instead of returning None/null.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1870435
Change-Id: I7bc4a7b65cb7a5c8b821ece737f73b1fe34555f7