livemerge: Separate job update and reporting
Mixing job updates and reporting makes the code more complicated and
hard to follow. This lead to bugs like not reporting jobs if the drive
was not found, and makes it hard to avoid reporting untracked jobs.

Extract the update part to _update_jobs(), which updates all jobs and
does not return anything. query_jobs() calls it and return a dict of
jobs info for engine.

We could create separate public API, and make callers call update() and
query() separately, but callers always want to get the job info dict and
update in the same time, so keeping the public API as is.

With this change we report jobs even if the drive was not found. We
never expect this case, but not reporting the job may confuse engine.

Change-Id: I37c0db0420d6a2f9c3dc983002f720813d48c0e2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>