livemerge: Introduce Job.state
Add Job.state tracking these states:

- COMMIT: The libvirt block commit job was started. This is the starting
  state when we start tracking a job.

- CLEANUP: Cleanup phase has started after the libvirt block commit job
  has gone, or reported as ready for pivot.

The new "state" variable is replacing the "gone" variable. If the
libvirt job is gone, we switch to COMMIT state and cleanup the job
without a pivot.

For active merge, we switch to CLEANUP state when libvirt reports that
the job is ready for pivot. We need to check if the libvirt job is ready
before starting a cleanup thread.

A nice side effect of this change is fixing a bug when reporting default
live info (cur=0, end=0) after successful pivot. Previous code tried to
get block job info and got empty dict, so default live info was
reported. Now we don't query libvirt job info after moving to CLEANUP
state so we continue to report the last live info.

Change-Id: I2473fa847dddc94e7c364803822313854439785b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>