livemerge: Eliminate fragile drive lookup
We use to lookup the drive object for every job in every _update_jobs()
iteration. Since drive lookup uses the job disk PDIV, it is expected to
fail after a pivot, when the drive volumeID changes from the top volume
to the base volume. We have a fallback to search the base volume that
probably works in runtime, but fail in the tests.

Eliminate this issue and simplify the code by moving drive lookup to
_start_cleanup_thread(). This always happens *before* pivot, so the
drive lookup failure is not possible. Since the lookup is simple now,
remove _lookup_drive().

Change-Id: I4c8bdc1183026d8db8ef7e9b43258243496490a0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>