livemerge: Move logging and state change to _start_cleanup()
We repeated the "Starting cleanup ..." log 4 times. Logging in
_start_cleaup() make it simpler.

The log about retrying cleanup is not really needed since we already log
the error at the end of the failing cleanup.

We set the job state to CLEANUP twice. Is it nicer if the state is
change in _start_cleanup().

Change-Id: Ib0006b8dfa98f120533611132987fc9a4fe6a4b0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>