livemerge: Improve error handling if updating a job fails
Previously if we had multiple tracked jobs, for example when deleting
snapshots with multiple disks, and updating one job raised we aborted
the update. Now failure to update a job will log an exception and
continue to update the rest of the jobs.

I did not have this issue, but looking at the simple updated loop in
query_jobs(), it was very clear that it is missing error handling.

Change-Id: I9f76983eb3efd530b08c75293db09cb1b8c14af8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>