vm: Add drive extend callback
Add a callback argument to extendDriveVolume(). If the callback is set,
it will be called when the extend completes. If the extend failed the
callback will be called with the error.

We will use this callback to start commit after the base volume was
extended.

We don't have any tests for this code, but this code is invoked from
live merge tests. The normal flow when callback is not set is tested now
by current livemerge tests. The flow with a callback will be tested once
we start using the callback.

Changed based on Amit patch:
https://gerrit.ovirt.org/c/vdsm/+/112835

Change-Id: Ib15b20e2451c790c08854ea9a5982bc5b3522d07
Signed-off-by: Nir Soffer <nsoffer@redhat.com>