livemerge: Fix job persistence
In the past persisting the jobs when adding or removing a job was
enough, since job metadata did not contain any state. Now that job keeps
internal state (EXTEND, COMMIT, CLEANUP) we must persist the job when we
change the state.

Persisting is tricky since vdsm may be killed at any point, and we want
to be able to recover. We always persist right after changing the job
state but before starting the actual operation. This way if vdsm is
killed, we should be able to recover.

For INIT state, the job is not persisted, so if vdsm is killed the job
does not exists and nothing has to be done.

For EXTEND state, if vdsm is killed before extend was started or
completed, we will treat the extend as a timeout and fail the job. With
more work we would be able to check if extend completed and start the
commit, or retry the extend.

For COMMIT state, if vdsm is killed before starting the commit we will
detect that the job does not exist in libvirt, and cleanup the job
without making any changes. With more work we can start the commit.

For CLEANUP state, if vdsm is killed before we started the cleanup, we
will detect that no cleanup thread exist, and start a new cleanup.

The tests use now parse_jobs(vm) helper to verify that jobs were
persisted with right state.

Change-Id: I7b0c87f2ae434b3cc6b856e8c36a5db71eb6ddc3
Bug-Url: https://bugzilla.redhat.com/1796415
Bug-Url: https://bugzilla.redhat.com/1796124
Signed-off-by: Nir Soffer <nsoffer@redhat.com>