storage: add LeaseMetadata to Lease.create
Add a metadata field to lease creation. It will be used
to store initial metadata on the lease's LVB.

Public fields will be sent by the engine upon lease creation:
generation, job_id and type.
The rest of the fields can only be set internally in vdsm.

Example for data present on LVB:
{
  'generation': 0,
  'job_status': 'PENDING',
  'job_id': '20e62244-19ad-46b7-afcc-a5fe2007a259',
  'type': 'job',
  'host_hardware_id': '6e11cad3-2214-4e1a-aadc-e050accfe3c0',
  'created': 1614073237,
  'modified': 1614073237
}

Change-Id: I594023a059aa3e56b3b6cff25fc1609e50dc7477
Bug-Url: https://bugzilla.redhat.com/1906074
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>