vm: Call setMetadata() and metadata() correctly
The flag argument is keyward argument, can seen in libvirt.py:

    def metadata(self, type, uri, flags=0):

    def setMetadata(self, type, metadata, key, uri, flags=0):

But the fake implementation was using:

    def metadata(self, type, uri, flags):

    def setMetadata(self, type, metadata, key, uri, flags):

So all code calling these methods was calling the function in the wrong
way. This tends to work since libvirt is unlikely to add new variables
or change the order, but it is bad practice.

Fix the wrong fakes and update the code to use the right way.

Change-Id: I3117490967731c003e3906ad6f76af35f572b6c7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>