tests: Fix flags handling
We stored VIR_DOMAIN_BLOCK_JOB_TYPE_ACTIVE_COMMIT for block job,
instead of VIR_DOMAIN_BLOCK_JOB_TYPE_COMMIT. The current code does not
check the type, so this issue was hidden.

When checking flags in blockJobAbort, we use == instead of testing the
flag bit. This works since this is the only flag used, but wrong.

Also update the test to verify job type.

Change-Id: Ife40ceb96651be5370d7f0b20dfcc11f85414617
Signed-off-by: Nir Soffer <nsoffer@redhat.com>