backup.py: skip disks validation when redefining a checkpoint
Checkpoint redefinition uses the backup configuration in order
to generate the checkpoint XML for libvirt. When creating the backup
configuration there is a validation for having disks in the backup
since there is no reason to start a backup without disks.

For checkpoint redefinition, there might be a case when there is a
checkpoint that should be redefined but it doesn't contain disks, for
example when the backed-up disks removed/detached from the VM.

The validation for the existence of the disks is now skipped in case of
using the backup configuration to redefine a checkpoint, also excluding
the <disks> element from the checkpoint XML.

Change-Id: Ic49f0df39f24f6952acd8d16cbbff8e118a1d605
Bug-Url: https://bugzilla.redhat.com/1925099
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>