storage: implement Lease.status
Lease.status already existed but did nothing. This patch implements the
method to be utilized by jobs using lease metadata.

It will be used to poll the job status when engine loses access to the
host that originally started the job.

The returned dict will be used to tell the status of the job.
If the owners list empty, engine will assume the job is still running
because the lease is still held.

If the owners list is empty, the generation will be used to determine
whether the job completed successfully or not.

Change-Id: Id0db704c060318501a35dc30d829b9838310b3f1
Bug-Url: https://bugzilla.redhat.com/1906074
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>