tests: Fix fake getVolumeInfo(), getVolumeSize()
The real API returns sizes as strings. This is leftover from the time
vdsm had xmlrpc API, which does not support big integers. Unfortunately
this was "fixed" internally by using strings, and when xmlrpc was
replaced by jsonrpc, the bad return value was not fixed.

Change getVolumeInfo() and getVolumeSize() to behave like the real API,
so the test will fail if code is using sizes as integer.

Change-Id: I00b7a9f6c963814ee2d8b08324215ab490c814c6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>