tests: fix test_all_channels_extra_domain() in vmxml
The test (added in commit b8d50777) is broken and it looks like it never
really worked. It is comparing for non-equality of list of strings
(channel names) and list of tuples (result of
DomainDescriptor.all_channels()). It seems that the original intent was
to make sure that all_channels() really returns all channels properly
and not just guest agent channels. So let's fix it this way.

At the time the test was added the function all_channels() also listed a
spice channel. But since commit 2a786fb this is no longer true and spice
channel is not returned. To remedy the situation new channel is added to
test domain XML.

Change-Id: I6d343fa37720d821983ee58e49a853cd1d669ee0
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>