qga: use libvirt events to track channel state
We can use events from libvirt to track the state of qemu-ga channel.
Libvirt reports when the agent connects to/disconnects from the virtio
channel. This is pretty useful because we're not in complete darkness
anymore and gives us at least a clue if the agent is there or not.
Obviously the agent can still fail to respond to our commands either
because it is stuck or because it is performing another command, but
most of our issues stemmed from the fact we didn't have a clue if agent
is there and were blindly trying to run commands (hoping for success).

This also fixes an issue that we were having ever since we switched to
libvirt interface. For VMs without the guest agent libvirt was spamming
journal with messages about agent being unavailable.

Change-Id: I3a79b21a93cf0b6c874b04e7305baad5757f3a94
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>