virt: Add error code for VIR_ERR_OPERATION_FAILED
VIR_ERR_OPERATION_FAILED libvirt error, when occurs during VM migration
in particular, is reported to the Engine as a generic 'migrateErr'. This
patch adds a separate 'migOperationErr' code for this error and passes
the accompanying message to the Engine, so it will be able to describe
the situation to the user.

Change-Id: Iadf377b6b1d15e26714608848ab6130a3ae886d5
Bug-Url: https://bugzilla.redhat.com/1717411
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>