virt: add flag for enabling/disabling ballooning
Even though we want to prevent certain VMs from ballooning the memory,
we still want to keep the balloon device in such VMs because that's how
we get memory statistics from the guest. So instead of removing balloon
device (and losing statistics) it makes sense to carry a flag whether
ballooning should be enabled or disabled. We read the initial value from
metadata key 'ballooningEnabled' (or default to True if not present) and
pass it over to MOM. We also ignore any balloon target changes for VMs
that have ballooning disabled.

Bug-Url: https://bugzilla.redhat.com/1917718
Change-Id: I402a37b2645e807350f97a5583b495af5840c039
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>