tests: test adding a lease fails when index is full
Coverage report revealed missing test for a case when index runs out
of space.

The simplest test would be to keep adding leases using public method
LeasesVolume.add() until the index fills up. However since we use
temporary file backend to test leases volume this operation is slow.

To speed up the test we can add a memory backend which will serve
for testing purposes only and will be placed alongside the existing
backends in xleases.py so it is not hidden.

The memory backend is using io.BytesIO to simulate a file.

Adding a new fixture that can be parametrized for tests that require
non default values for memory backend (alignment, block_size, size).

The actual test uses this fixture to create a backend with limited size
so that only one lease can be added so next attempt will raise NoSpace.
If the parameters are not used reasonable defaults are applied.

Another limitation is xleases.format_index() which had hardcoded
amount of records. Adding a parameter so this can be controlled.
Without it the memory backend always gets rewritten and expanded to
index max size silently because io.BytesIO is dynamically allocated
rendering backend size parameter useless.

Another approach would be creating a backend with zero size but that
does not work because LeasesVolume checks for storage format and fails
if the format is not met. Alternatively using sufficient size but with
zeroes only results in the same failure.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Change-Id: Ie802098fbb75ad971fcec33e90ba5083f31b4510