virt: add "managed" metadata key to drives
The "managed" key will indicate whether the drive is already managed
(Managed Block Storage) or needs to be managed by vdsm.

This will be used to avoid device appropriation when start a VM with
managed disks attached, as they are appropriated before the VM started,
when they are connected to host.

Example of the how the metadata looks with this field:
<ovirt-vm:device devtype="disk" name="sdb">
    <ovirt-vm:GUID>3600a098038304479363f4c4870455167</ovirt-vm:GUID>
    <ovirt-vm:imageID>3600a098038304479363f4c4870455167</ovirt-vm:imageID>
    <ovirt-vm:managed type="bool">true</ovirt-vm:managed>
</ovirt-vm:device>

Change-Id: Iedb7834da9c2239eb268ae14a38db1fc25ad0a28
Bug-Url: https://bugzilla.redhat.com/1906074
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>