net: Properly remove old source routes
There were some unnoticed cases in which outdated
source routes were not removed. This was uncovered by
testing vdsm with nmstate 1.0. To prevent this issue
add method to remove source route when it is relevant.

The most important flows are switch from static to
dynamic or vice versa, removal of network (kernel has
removed routes but not rules), change of gateway.

At the same time optimize addition of source routes.

Change-Id: Ia9485efb9b29afee7c298f6fa4ff320d87e99216
Signed-off-by: Ales Musil <amusil@redhat.com>