backup.py: remove parent_checkpoint_id
Since all the checkpoint are active on the VM there is
no dependency between them. When a checkpoint is used
to start a backup from that point in time, there is no
need to link it to the previous checkpoint that was taken
and chain them.

Chaining the checkpoint is done by specifying the parent
checkpoint ID, the element can be dropped now from the
checkpoint XML and the parent_checkpoint_id has no other
use and can be removed.

This patch removes the parent_checkpoint_id from the
checkpoint XML and the  backup config parameters with
the validation related to that field.
Also, fix the tests accordingly.

Change-Id: Id8ab9e1489e65edde32be1a5d009af61805157e1
Bug-Url: https://bugzilla.redhat.com/1950752
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>