livemerge: Fix wrong name
Vm.refreshDriveVolume() was renamed recently to
Vm.refresh_drive_volume(). When the live merge code moved to livemerge
module in:

commit a8c3d80af92c2faa0713603d3d1eef2e5418249b

    vm: Move live merge code to livemerge module

The old name was used by mistake. Unfortunately this flow is not covered
by the tests since it is hard to create a test vm with this
configuration, and the edge case of deleting snapshot on a raw volume
after extending the disk was not tested.

Change-Id: Ie83169ffed4109b6417768750e149dcabaf0a2b8
Bug-Url: https://bugzilla.redhat.com/1941311
Reported-by: Evelina Shames <eshames@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>