backup.py: add bitmap name for disks when redefining a checkpoint
Libvirt changed the default behavior after dropping the need to
redefine with the VM domain XML. Before that, when a checkpoint
redefined, if the bitmap name wasn't mentioned for the disk it was
taken from the domaincheckpoint name.

Now, Libvirt changed that behavior so we should provide the bitmap
name for each disk in the backup.

Change-Id: Id83beaac641daf094882bbf325083a9723769932
Bug-Url: https://bugzilla.redhat.com/1941593
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>
(cherry picked from commit 9d33896c066f9f08049067da9f81946eab800ee3)