virt: optionally forbid reading of empty external data
An option was added to allow/forbid reading empty data (zero-sized file
or empty directory). In case the requirement is broken
ExternalDataFailure exception is raised. This can slightly mitigate
situations where data are damaged (truncated) due to host crashes as it
prevents sending the empty data to Engine.

The default (affecting both NVRAM and TPM) is that empty data are not
allowed.

Bug-Url: https://bugzilla.redhat.com/1943141
Change-Id: I4c6de487cd324addbbc260e606b2aa834f7a377b
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>