live-merge: Prepare for extend retries
When extend does not complete, we need to retry the operation. To
implement this we need to keep more info about the extend. Replace
"extend_started" with "extend" dict. This make it easy to add more
fields as needed, minimize the size of the metadata when we are not in
EXTEND state, and simplify the code to manage the metadata.

Here is an example metadata with this change:

{
    "5c0a3978-d3c0-4acb-84ee-a52bf0088878": {
        "bandwidth": 0,
        "baseVolume": "2939852e-187c-48b4-b57a-1f8ea5bc94f8",
        "disk": {
            "poolID": "84fab540-fbf2-11ea-a568-5254002fb5cc",
            "domainID": "c0b55558-5038-4a2b-bc96-47aff28d0218",
            "imageID": "222d9718-f2d9-4b95-bd7f-e8dfe565cf56",
            "volumeID": "6c0f1c64-90ac-4a68-839d-5a47af1d3890"
        },
        "drive": "sda",
        "extend": {
            "started": 15051.875996275
        },
        "jobID": "5c0a3978-d3c0-4acb-84ee-a52bf0088878",
        "state": "EXTEND",
        "topVolume": "6c0f1c64-90ac-4a68-839d-5a47af1d3890"
    }
}

Change-Id: I3c953d5f76cae3a3e194f6fc1650cd8d18c41668
Signed-off-by: Nir Soffer <nsoffer@redhat.com>