virt: always report fake guest application info
VDSM reports fake entries in guest application list for qemu-ga and for
linux kernel. These were however reported only if oVirt guest agent was
not running or if the application list did not contain entry for
qemu-ga.

Because the entry reported for the qemu-ga reported by oVirt guest agent
does not contain version of qemu-ga this behavior is no longer favorable
since Engine needs to know the version to be able to give indication to
the user for updating tools. So instead, we now report the fake entries
always.

To avoid some duplicates in the output, entries "QEMU guest agent"
(on Windows) and "kernel-<version>" (on Linux) produced by oVirt guest
agent are removed.

Change-Id: Iead4caa5183bbe402009c8d4c900b852a121becb
Bug-Url: https://bugzilla.redhat.com/1936298
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>