snapshot: fix teardown handling
When we teardown the snapshot operation, after the pivot and updating
the drives parameter, we shouldn't fail the snapshot job. This case may
happen when the host loses its connection to the storage, on the second
attempt, some values such as volume ID might be missing. But the
snapshot was created successfully and is usable.

Change-Id: I4e211501e2344dcce3e836666351b7bf37fd1edc
Bug-Url: https://bugzilla.redhat.com/1946193
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>