lvm: Do not prompt when wiping signatures
LVM changed the behavior when wiping signatures during lvcreate, and now
it requires manual confirmation. This breaks lvcreate calls in vdsm if
the lv has lefover from previous usage of the lv.

Add --yes to lvcreate to avoid the confirmation.

More work may be needed to updated the tests. I post this as to check if
this unblocks automated tests.

Change-Id: I1fb5d9deccb08939d9a27628156d0d34084e6178
Bug-Url: https://bugzilla.redhat.com/1946199
Related-to: https://bugzilla.redhat.com/1894692
Signed-off-by: Nir Soffer <nsoffer@redhat.com>