udev: add functions for managed devices
This patch adds specific functions to handle udev operations for
MBS devices
* add_managed_udev_rule
* trigger_managed_udev_rule
* remove_managed_udev_rule

We need separate functions as appropriateDevice expects a "thiefId"
which is essentially a VM ID that owns the udev rule. For MBS devices,
we do not have a VM as we run the rule before starting the VM or in
flows where a VM is not involved like copy_data.

Change-Id: I6134149f289514019fe46f849436fce96c0aa8f2
Bug-Url: https://bugzilla.redhat.com/1906074
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>