livemerge: Always retry failed pivot
In commit bbb7e6f8ded52025308727dc3f29d514ea175bfc

    vm: Stop live merge cleanup attempts for unrecoverable pivot error

We changed error handling for unexpected libvirt errors when trying to
pivot. Based on advice from libvirt folks, we assumed that unexpected
error is not recoverable, and we aborted the merge job immediately.

There are major issues with that commit:

- The assumption that unexpected error is no recoverable is not always
  correct, as we can see in bug[1]. In this bug, qemu block jobs state
  is flipping between "ready" and "standby", and retrying the pivot
  several times is very likely to complete successfully.

- We aborted the merge job without aborting the libvirt block job! This
  leave the merge in unrecoverable state; because the libvirt job is
  still running, starting new merge will fail immediately. And because
  the top volume is marked as ILLEGAL, the VM will not run after you
  stop the VM.

Fixing this issue requires 2 changes:

1. Always retry cleanup after pivot errors
2. Limit the number of retries and abort the libvirt block job.

This change implement the first part, which is easy, fixes the
regression, and should be enough to mitigate bug[1] in case it is not
fixed soon in qemu.

The second part of the fix is more complicated and will be done later,
since we have more urgent work to do now.

[1] https://bugzilla.redhat.com/1945635

Change-Id: I1fcc9893a5ad996248afcdc5877616656ec3adcc
Bug-Url: https://bugzilla.redhat.com/1945675
Signed-off-by: Nir Soffer <nsoffer@redhat.com>