livemerge: Keep pivot state in job
Previously we checked if a job is ready before retrying cleanup. The
motivation was to handle unexpected case when libvirt job moves from
ready state to non-ready state. In this case we create the cleanup
thread with doPivot=False. This cannot work because the cleanup thread
does not complete the block job if doPivot=False, and we may leave a
running block job, which cannot be recovered.

Fortunately, libvirt block job never moves from "ready" state after
ready state was reported, so we never try to cleanup active commit with
doPivot=False.

Change to check if job is ready only in COMMIT state. When an active
commit job is ready for pivot, we set job.pivot to True. The next
cleanup attempt will try to pivot without checking libvirt state again.
The pivot flag is persisted in the VM metadata, so after recovery we
will try to pivot again.

Change-Id: I4b968be857929a911e0c361f6e9408ebc36ca6ea
Bug-Url: https://bugzilla.redhat.com/1945675
Signed-off-by: Nir Soffer <nsoffer@redhat.com>