livemerge: Fix cleanup if active commit job has gone
If active commit block job is canceled outside of vdsm when a job is in
CLEANUP state, for example using virsh, vdsm will be stuck in endless
pivot retry loop.

In the past vdsm could detect this case since it checked job live info
on every call to queryBlockJobs(). Since we introduced job states, the
check is done only in COMMIT state in _update_commit().

Canceling job outside of vdsm should not happen, but this may be a good
way to cancel jobs that vdsm cannot complete. For example if pivot fail
because of libvirt or qemu bug[1]. If this happens, we want vdsm to
detect the situation and abort the merge.

Fix by checking if libvirt block job exists also in CLEANUP state. If
the job has gone, we change job.pivot to False, and retry the cleanup.

[1] https://bugzilla.redhat.com/1945635

Change-Id: Icc0a10842aa0065765e71e450a1a3963b01ced41
Bug-Url: https://bugzilla.redhat.com/1945675
Signed-off-by: Nir Soffer <nsoffer@redhat.com>