livemerge: Do not retry extend on errors
The motivation for trying extend on errors was to recover faster from
errors. However if an extend fails because of a temporary issue, like
inaccessible storage, waiting before retrying may increase the chance to
recover from the error.

Retrying extend can also lead to extend storm, when extend request fail
quickly and starts another extend attempt, which will fail again quickly
and start the next attempt. We may use all retries in few seconds which
is not useful.

Fixed by removing the retry on extend errors. The next retry will start
on the next timeout. This is also much simpler to reason about.

Change-Id: I33528bbcfe897e9454e563419476c4da9615f81d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>